[sensors-reset] Migrate tests
