--- sensors_reset/tests/specs/incomplete_task_three_reset.yml
@@ -1,13 +1,0 @@
-initialization:
-  clone-from: https://github.com/git-mastery/gm-sensors
-  steps:
-    - type: commit
-      empty: true
-      message: Empty commit
-      id: start
-    - type: reset
-      revision: "5950516"
-      mode: hard
-    - type: reset
-      revision: "f900a07"
-      mode: mixed
