--- sensors_reset/tests/specs/incorrect_head_commit.yml
@@ -1,19 +1,0 @@
-initialization:
-  clone-from: https://github.com/git-mastery/gm-sensors
-  steps:
-    - type: commit
-      empty: true
-      message: Empty commit
-      id: start
-    - type: reset
-      revision: "5950516"
-      mode: hard
-    - type: reset
-      revision: "f900a07"
-      mode: mixed
-    - type: reset
-      revision: "c47b187"
-      mode: soft
-    - type: reset
-      revision: "e45a4fc"
-      mode: soft
