--- sensors_reset/tests/specs/no_changes.yml
@@ -1,7 +1,0 @@
-initialization:
-  clone-from: https://github.com/git-mastery/gm-sensors
-  steps:
-    - type: commit
-      empty: true
-      message: Empty commit
-      id: start
